backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
stream-chat==4.15.0
pyjwt==2.8.0
